Fix inconsistent handling of dynamic sizes
__ESBMC_alloc_size was an unsigned long (which has a different
size in 32 bits and 64 bits archs) and we were assigning various
objects with different types to it, in particular, int and pointer_type.

It was OK in 32 bits archs because these types have the same bit-width
but hell broke loose when we set 64 bits to be the default arch in
ESBMC.

+++ src/clang-c-frontend/clang_c_language.cpp
@@ -275,7 +275,7 @@ __attribute__((annotate("__ESBMC_inf_size")))
 _Bool __ESBMC_is_dynamic[1];
 
 __attribute__((annotate("__ESBMC_inf_size")))
-unsigned long __ESBMC_alloc_size[1];
+unsigned __ESBMC_alloc_size[1];
 
 _Bool __ESBMC_is_little_endian();
